SNO+: Account for roll over in cmos rate polling
Also change the internal tracking of the CMOS rate to use
uint32_t rather than unsigned long

+++ Source/Objects/Custom Hardware/SNO+/Fec32/ORFec32Model.h
@@ -376,7 +376,7 @@ typedef struct Fec32CmosShiftReg{
 - (void) scan:(SEL)aResumeSelectorInGuardian; 
 - (void) scanWorkingSlot;
 - (BOOL) readCMOSCounts:(BOOL)calcRates channelMask:(unsigned long) aChannelMask;
-- (BOOL) processCMOSCounts:(unsigned long*)rates calcRates:(BOOL)aCalcRates withChannelMask:(unsigned long) aChannelMask;
+- (BOOL) processCMOSCounts:(uint32_t*)rates calcRates:(BOOL)aCalcRates withChannelMask:(unsigned long) aChannelMask;
 - (unsigned long) channelsWithCMOSRateHigherThan:(unsigned long)cmosRateLimit;
 - (unsigned long) channelsWithErrorCMOSRate;
